Refactor formula compiler into modular, maintainable ES module structure
